requirements.txt
requests==2.25.1
urllib3==1.26.5
pymongo==3.11.4
python_dateutil==2.8.2
